backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
langchain==0.1.16
langchain-community==0.0.34
langchain-huggingface==0.0.1
chromadb==0.5.0
sentence-transformers==2.7.0
pypdf==4.2.0
mlflow==2.13.0
langchain-groq==0.1.3
python-dotenv==1.0.1
